third_party/blink/web_tests/virtual/customizable-select-disabled/html5lib/generated/run-webkit02-write-expected.txt

CONSOLE WARNING: A xh<optgroup tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
../resources/webkit02.dat:
3
12
16
17
18

Test 3 of 35 in ../resources/webkit02.dat failed. Input:
<p id="status"><noscript><strong>A</strong></noscript><span>B</span></p>
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       id="status"
|       <noscript>
|         "<strong>A</strong>"
|       <span>
|         "B"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       id="status"
|       <noscript>
|         <strong>
|           "A"
|       <span>
|         "B"

Test 12 of 35 in ../resources/webkit02.dat failed. Input:
<b><em><dcell><postfield><postfield><postfield><postfield><missing_glyph><missing_glyph><missing_glyph><missing_glyph><hkern><aside></b></em>
Got:
| <b>
|   <em>
|     <dcell>
|       <postfield>
|         <postfield>
|           <postfield>
|             <postfield>
|               <missing_glyph>
|                 <missing_glyph>
|                   <missing_glyph>
|                     <missing_glyph>
|                       <hkern>
| <aside>
|   <em>
|     <b>
Expected:
| <b>
|   <em>
|     <dcell>
|       <postfield>
|         <postfield>
|           <postfield>
|             <postfield>
|               <missing_glyph>
|                 <missing_glyph>
|                   <missing_glyph>
|                     <missing_glyph>
|                       <hkern>
| <aside>
|   <b>

Test 16 of 35 in ../resources/webkit02.dat failed. Input:
<b><em><foo><foo><foo><aside></b></em>
Got:
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|             <foo>
|     <aside>
|       <em>
|         <b>
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|             <foo>
|     <aside>
|       <b>

Test 17 of 35 in ../resources/webkit02.dat failed. Input:
<b><em><foo><foo><foo><foo><foo><foo><foo><foo><foo><foo><aside></b></em>
Got:
| <b>
|   <em>
|     <foo>
|       <foo>
|         <foo>
|           <foo>
|             <foo>
|               <foo>
|                 <foo>
|                   <foo>
|                     <foo>
|                       <foo>
| <aside>
|   <em>
|     <b>
Expected:
| <b>
|   <em>
|     <foo>
|       <foo>
|         <foo>
|           <foo>
|             <foo>
|               <foo>
|                 <foo>
|                   <foo>
|                     <foo>
|                       <foo>
| <aside>
|   <b>

Test 18 of 35 in ../resources/webkit02.dat failed. Input:
<b><em><foo><foob><foob><foob><foob><fooc><fooc><fooc><fooc><food><aside></b></em>
Got:
| <b>
|   <em>
|     <foo>
|       <foob>
|         <foob>
|           <foob>
|             <foob>
|               <fooc>
|                 <fooc>
|                   <fooc>
|                     <fooc>
|                       <food>
| <aside>
|   <em>
|     <b>
Expected:
| <b>
|   <em>
|     <foo>
|       <foob>
|         <foob>
|           <foob>
|             <foob>
|               <fooc>
|                 <fooc>
|                   <fooc>
|                     <fooc>
|                       <food>
| <aside>
|   <b>
